+++ handler/core/validator.go
@@ -4,17 +4,19 @@ import (
 	"github.com/ory-am/fosite"
 	"github.com/pkg/errors"
 	"golang.org/x/net/context"
+	"fmt"
 )
 
 type CoreValidator struct {
 	CoreStrategy
 	CoreStorage
+	ScopeStrategy fosite.ScopeStrategy
 }
 
-func (c *CoreValidator) ValidateToken(ctx context.Context, token string, tokenType fosite.TokenType, accessRequest fosite.AccessRequester) error {
+func (c *CoreValidator) ValidateToken(ctx context.Context, token string, tokenType fosite.TokenType, accessRequest fosite.AccessRequester, scopes []string) error {
 	switch tokenType {
 	case fosite.AccessToken:
-		return c.validateAccessToken(ctx, token, accessRequest)
+		return c.validateAccessToken(ctx, token, accessRequest, scopes)
 	case fosite.RefreshToken:
 		return c.validateRefreshToken(ctx, token, accessRequest)
 	case fosite.AuthorizeCode:
@@ -24,16 +26,22 @@ func (c *CoreValidator) ValidateToken(ctx context.Context, token string, tokenTy
 	}
 }
 
-func (c *CoreValidator) validateAccessToken(ctx context.Context, token string, accessRequest fosite.AccessRequester) error {
+func (c *CoreValidator) validateAccessToken(ctx context.Context, token string, accessRequest fosite.AccessRequester, scopes []string) error {
 	sig := c.CoreStrategy.AccessTokenSignature(token)
-	if or, err := c.CoreStorage.GetAccessTokenSession(ctx, sig, accessRequest.GetSession()); err != nil {
+	or, err := c.CoreStorage.GetAccessTokenSession(ctx, sig, accessRequest.GetSession())
+	if err != nil {
 		return errors.Wrap(fosite.ErrRequestUnauthorized, err.Error())
 	} else if err := c.CoreStrategy.ValidateAccessToken(ctx, or, token); err != nil {
 		return errors.Wrap(fosite.ErrRequestUnauthorized, err.Error())
-	} else {
-		accessRequest.Merge(or)
 	}
 
+	for _, scope := range scopes {
+		if !c.ScopeStrategy(accessRequest.GetGrantedScopes(), scope) {
+			return errors.Wrap(fosite.ErrRequestForbidden, fmt.Sprintf("scope %s was not granted", scope))
+		}
+	}
+
+	accessRequest.Merge(or)
 	return nil
 }