+++ validator_test.go
@@ -46,7 +46,7 @@ func TestValidate(t *testing.T) {
 			scopes:      []string{"foo"},
 			setup: func() {
 				f.TokenValidators = TokenValidators{validator}
-				validator.EXPECT().ValidateToken(nil, "some-token", gomock.Any(), gomock.Any()).Return(ErrUnknownRequest)
+				validator.EXPECT().ValidateToken(nil, "some-token", gomock.Any(), gomock.Any(), gomock.Any()).Return(ErrUnknownRequest)
 			},
 			expectErr: ErrRequestUnauthorized,
 		},
@@ -54,24 +54,14 @@ func TestValidate(t *testing.T) {
 			description: "should fail",
 			scopes:      []string{"foo"},
 			setup: func() {
-				validator.EXPECT().ValidateToken(nil, "some-token", gomock.Any(), gomock.Any()).Return(ErrInvalidClient)
+				validator.EXPECT().ValidateToken(nil, "some-token", gomock.Any(), gomock.Any(), gomock.Any()).Return(ErrInvalidClient)
 			},
 			expectErr: ErrInvalidClient,
 		},
 		{
-			description: "should fail",
-			scopes:      []string{"foo"},
-			setup: func() {
-				validator.EXPECT().ValidateToken(nil, "some-token", gomock.Any(), gomock.Any()).Do(func(ctx context.Context, _ string, _ TokenType, accessRequest AccessRequester) {
-					accessRequest.(*AccessRequest).GrantedScopes = []string{"bar"}
-				}).Return(nil)
-			},
-			expectErr: ErrRequestForbidden,
-		},
-		{
 			description: "should pass",
 			setup: func() {
-				validator.EXPECT().ValidateToken(nil, "some-token", gomock.Any(), gomock.Any()).Do(func(ctx context.Context, _ string, _ TokenType, accessRequest AccessRequester) {
+				validator.EXPECT().ValidateToken(nil, "some-token", gomock.Any(), gomock.Any(), gomock.Any()).Do(func(ctx context.Context, _ string, _ TokenType, accessRequest AccessRequester, _ []string) {
 					accessRequest.(*AccessRequest).GrantedScopes = []string{"bar"}
 				}).Return(nil)
 			},
@@ -80,7 +70,7 @@ func TestValidate(t *testing.T) {
 			description: "should pass",
 			scopes:      []string{"bar"},
 			setup: func() {
-				validator.EXPECT().ValidateToken(nil, "some-token", gomock.Any(), gomock.Any()).Do(func(ctx context.Context, _ string, _ TokenType, accessRequest AccessRequester) {
+				validator.EXPECT().ValidateToken(nil, "some-token", gomock.Any(), gomock.Any(), gomock.Any()).Do(func(ctx context.Context, _ string, _ TokenType, accessRequest AccessRequester, _ []string) {
 					accessRequest.(*AccessRequest).GrantedScopes = []string{"bar"}
 				}).Return(nil)
 			},